backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai>=0.8.3
langchain-google-genai
langchain-chroma
chromadb
google-cloud-texttospeech
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

PyPDF2
python-docx
openpyxl
Pillow
beautifulsoup4
requests
lxml
lxml
tzdata
langchain-text-splitters
duckduckgo-search
